Fix GCC and Release warnings

+++ CesiumGltf/include/CesiumGltf/AccessorUtility.h
@@ -242,13 +242,17 @@ struct IndicesForFaceFromAccessor {
 
     if (primitiveMode == MeshPrimitive::Mode::TRIANGLE_FAN) {
       result[0] = value[0];
-      result[1] = firstIndex < value.size() ? value[firstIndex] : -1;
-      result[2] = firstIndex + 1 < value.size() ? value[firstIndex + 1] : -1;
+      result[1] = firstIndex < value.size()
+                      ? static_cast<int64_t>(value[firstIndex])
+                      : -1;
+      result[2] = firstIndex + 1 < value.size()
+                      ? static_cast<int64_t>(value[firstIndex + 1])
+                      : -1;
     } else {
       for (int64_t i = 0; i < 3; i++) {
         int64_t index = firstIndex + i;
         result[static_cast<size_t>(i)] =
-            index < value.size() ? value[index] : -1;
+            index < value.size() ? static_cast<int64_t>(value[index]) : -1;
       }
     }
